# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NexaChat AI API
  version: 2.0.0
  description: Owner-isolated multimodal AI workspace API. See API.md for the full behavioral contract.
servers:
  - url: http://localhost:5000
components:
  parameters:
    CsrfToken:
      in: header
      name: X-CSRF-Token
      required: true
      schema: { type: string }
      description: Session-bound token returned by GET /api/config.
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
    Conversation:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        persona: { type: string }
        model: { type: string }
        is_pinned: { type: boolean }
        is_archived: { type: boolean }
        message_count: { type: integer }
...
paths:
...
  /api/conversations:
    get:
      summary: List owned conversations
      responses:
        "200":
          description: Conversation list
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Conversation" } }
    post:
      summary: Create conversation
      parameters: [{ $ref: "#/components/parameters/CsrfToken" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                model: { type: string }
                persona: { type: string, enum: [general, coding, study, career, product] }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Conversation" }
        "403":
          description: CSRF rejected
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      summary: Delete all owned conversations
      parameters: [{ $ref: "#/components/parameters/CsrfToken" }]
      responses:
        "204": { description: Deleted }
  /api/conversations/{conversationId}:
    parameters:
      - in: path
        name: conversationId
        required: true
        schema: { type: integer }
    get:
      summary: Get conversation with messages
      responses:
        "200": { description: Conversation }
        "404": { description: Missing or unowned }
    patch:
      summary: Update title, model, persona, instructions, or pin
      parameters: [{ $ref: "#/components/parameters/CsrfToken" }]
      requestBody:
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200": { description: Updated }
    delete:
      summary: Delete conversation
      parameters: [{ $ref: "#/components/parameters/CsrfToken" }]
      responses:
        "204": { description: Deleted }
  /api/conversations/{conversationId}/messages:
    post:
      summary: Send message and receive SSE tokens
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - in: path
          name: conversationId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
                model: { type: string }
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream: { schema: { type: string } }
        "409":
          description: Current data requires a live-research plan
...